+++ Changelog.md
@@ -1,9 +1,10 @@
 # Next
 
 
-- **Breaking Change** Removes "multipartBody" from TargetType protocol and adds a "task" instead
-- **Breaking Change** Successful Responses that have no data with them are now being converted to .Success Results
-- Adds Download and Upload Task type support to Moya
+- **Breaking Change** Removes `multipartBody` from TargetType protocol and adds a `task` instead.
+- **Breaking Change** Successful Response instances that have no data with them are now being converted to `.Success` `Result`s.
+- Adds Download and Upload Task type support to Moya.
+- Corrects SwiftLint warnings.
 
 # 7.0.0