+++ Changelog.md
@@ -11,6 +11,10 @@
 - Added `supportsMultipart` to the `Method` type, which helps determine whether to use `multipart/form-data` encoding.
 - Added `PATCH` and `CONNECT` to the `Method` cases which support multipart encoding.
 
+# 7.0.1
+
+- Add support to swift 2.3
+
 # 7.0.0
 
 - **Breaking Change** Drops support for `RACSignal`.